move src/read_exact.rs -> src/io/read_exact.rs
